--- types/contract/contractPayment.d.ts
@@ -1,8 +1,0 @@
-import { PaymentGateway } from './../payment'
-
-export interface IPreparePaymentResponse {
-  contractId: number
-  paymentGateway: PaymentGateway
-  checkoutSessionId?: string
-  stripePublicKey?: string
-}
